// packages/examples/src/enums.ts
export enum DataTestId {
    Checkbox1 = "checkbox-1",
    Checkbox2 = "checkbox-2",
    Checkbox3 = "checkbox-3",
    ClassComponent = "class-component",
    ClassInput = "class-input",
    ClassSelect = "class-select",
    ClassSubmit = "class-submit",
    DynamicComponent = "dynamic-component",
    DynamicContent = "dynamic-content",
    Error = "error",
    Error1 = "error-1",
    Error2 = "error-2",
    Error3 = "error-3",
    FormController = "form-controller",
    FunctionalComponent = "functional-component",
    FunctionalInput = "functional-input",
    FuntionalSelect = "functional-select",
    FunctionalSubmit = "functional-submit",
    GivenName = "givenName",
    InputFieldRow1 = "input-field-row-1",
    InputFieldRow2 = "input-field-row-2",
    InvalidText = "invalid-text",
    Middlename = "middlename",
    Note = "note",
    Radio1 = "radio-1",
    Radio2 = "radio-2",
    Radio3 = "radio-3",
    Radio11 = "radio-1-1",
    Radio12 = "radio-1-2",
    Radio13 = "radio-1-3",
    Radio21 = "radio-2-1",
    Radio22 = "radio-2-2",
    Radio23 = "radio-2-3",
    Radio31 = "radio-3-1",
    Radio32 = "radio-3-2",
    Radio33 = "radio-3-3",
    RadioFieldRow1 = "radio-field-row-1",
    RadioFieldRow2 = "radio-field-row-2",
    RadioFieldRow3 = "radio-field-row-3",
    ReRender = "re-render",
    Reset = "reset",
    Salutation = "salutation",
    Select = "select",
    Select1 = "select-1",
    Select2 = "select-2",
    SelectFieldRow = "select-field-row",
    Submit = "submit",
    SubmitBottom = "submit-bottom",
    SubmitTop = "submit-top",
    Surname = "surname",
    Textarea = "textarea",
    TextareaFieldRow = "textarea-field-row",
    ValidText = "valid-text"
}

export enum TestingContent {
    Bond = "Bond",
    Description = "Description",
    James = "James",
    JamesBond = "James Bond",
    JamesJunior = "James Junior",
    GivenName = "Given name",
    CaptionCheckbox = "Checkbox",
    CaptionCheckbox1 = "Checkbox 1",
    CaptionCheckbox2 = "Checkbox 2",
    CaptionCheckbox3 = "Checkbox 3",
    CaptionOption1 = "Option 1",
    CaptionOption11 = "Option 1-1",
    CaptionOption12 = "Option 1-2",
    CaptionOption13 = "Option 1-3",
    CaptionOption2 = "Option 2",
    CaptionOption21 = "Option 2-1",
    CaptionOption22 = "Option 2-2",
    CaptionOption23 = "Option 2-3",
    CaptionOption3 = "Option 3",
    CaptionOption31 = "Option 3-1",
    CaptionOption32 = "Option 3-2",
    CaptionOption33 = "Option 3-3",
    InputGivenName = "Input a given name",
    InputSurname = "Input a surname",
    InputText = "Input a text",
    Pending = "pending...",
    Ronald = "Ronald",
    Salutation = "Salutation",
    Select = "Select",
    Select1 = "Select 1",
    Select2 = "Select 2",
    Surname = "Surname",
    ValueCheckbox1 = "checkbox-1",
    ValueCheckbox2 = "checkbox-2",
    ValueCheckbox3 = "checkbox-3",
    ValueOption1 = "option-1",
    ValueOption11 = "option-1-1",
    ValueOption12 = "option-1-2",
    ValueOption13 = "option-1-3",
    ValueOption2 = "option-2",
    ValueOption21 = "option-2-1",
    ValueOption22 = "option-2-2",
    ValueOption23 = "option-2-3",
    ValueOption3 = "option-3",
    ValueOption31 = "option-3-1",
    ValueOption32 = "option-3-2",
    ValueOption33 = "option-3-3"
}